requirements.txt
# Core dependencies
requests==2.32.5
python-magic==0.4.27

# PDF processing
PyMuPDF==1.26.7

# Image processing
Pillow==12.1.0
transformers==5.0.0

# OCR (uses the Tesseract binary installed via apt)
pytesseract==0.3.13

# Testing
pytest==9.0.2
pytest-cov==7.0.0

# Utilities
python-dotenv==1.2.1
tqdm==4.67.2